refactor(hypervisor): move main Hypervisor struct into top-level hypervisor module
