+++ hypervisor/src/hypervisor/partition.rs
@@ -8,6 +8,15 @@ use std::time::Duration;
 
 use a653rs::bindings::PortDirection;
 use a653rs::prelude::{OperatingMode, StartCondition};
+use anyhow::{anyhow, bail};
+use clone3::Clone3;
+use itertools::Itertools;
+use nix::mount::{mount, umount2, MntFlags, MsFlags};
+use nix::sys::socket::{self, bind, AddressFamily, SockFlag, SockType, UnixAddr};
+use nix::unistd::{chdir, close, pivot_root, setgid, setuid, Gid, Pid, Uid};
+use procfs::process::Process;
+use tempfile::{tempdir, TempDir};
+
 use a653rs_linux_core::cgroup::CGroup;
 use a653rs_linux_core::error::{
     ErrorLevel, LeveledResult, ResultExt, SystemError, TypedResult, TypedResultExt,
@@ -19,20 +28,13 @@ use a653rs_linux_core::ipc::{channel_pair, io_pair, IoReceiver, IoSender, IpcRec
 use a653rs_linux_core::partition::{PartitionConstants, SamplingConstant};
 use a653rs_linux_core::sampling::Sampling;
 use a653rs_linux_core::syscall::SYSCALL_SOCKET_PATH;
-use anyhow::{anyhow, bail};
-use clone3::Clone3;
-use itertools::Itertools;
-use nix::mount::{mount, umount2, MntFlags, MsFlags};
-use nix::sys::socket::{self, bind, AddressFamily, SockFlag, SockType, UnixAddr};
-use nix::unistd::{chdir, close, pivot_root, setgid, setuid, Gid, Pid, Uid};
-use procfs::process::Process;
-use tempfile::{tempdir, TempDir};
+
+use crate::hypervisor::config::Partition as PartitionConfig;
+use crate::hypervisor::SYSTEM_START_TIME;
+use crate::problem;
 
 use super::config::PosixSocket;
 use super::scheduler::{PartitionTimeWindow, Timeout};
-use crate::hypervisor::config::Partition as PartitionConfig;
-use crate::hypervisor::linux::SYSTEM_START_TIME;
-use crate::problem;
 
 #[derive(Debug, Clone, Copy)]
 pub enum TransitionAction {
